refactor(viewer): collapse url-loading state + dedup splitAlias
Two simplifications spotted in the post-feature review:

(1) The frontend was carrying a `landingState.proxyEnabled` boolean
    that existed only to be ternaried into the FileUpload's
    `urlLoading: 'proxy' | 'disabled'` prop. Have /api/v1/mode
    return the tri-state value directly — backend reports "proxy"
    or "disabled", static-site bootstrap continues to hardcode
    "direct". Drops the boolean and the translation, both
    script.js bootstraps end up symmetric at the FileUpload init.

(2) `splitAlias` was implemented twice — once in the static-site
    bootstrap, once inline (without the function name) inside the
    binary viewer's onLoadUrl handler. Both are the JS twin of
    Rust's split_alias in src/viewer/mod.rs. Move the function
    into landing.js as an export and import from both callers so
    the alias-grammar rules live in one place.

No behavior change.
